add----增加部分方法

+++ src/main/java/com/immediateactiongroup/issues/service/impl/UserServiceImpl.java
@@ -1,5 +1,7 @@
 package com.immediateactiongroup.issues.service.impl;
 
+import com.immediateactiongroup.issues.commons.exception.BusinessException;
+import com.immediateactiongroup.issues.commons.exception.ExceptionEnum;
 import com.immediateactiongroup.issues.dto.AddUserDTO;
 import com.immediateactiongroup.issues.dto.UserDTO;
 import com.immediateactiongroup.issues.model.Role;
@@ -7,6 +9,8 @@ import com.immediateactiongroup.issues.model.User;
 import com.immediateactiongroup.issues.model.repository.RoleRepository;
 import com.immediateactiongroup.issues.model.repository.UserRepository;
 import com.immediateactiongroup.issues.service.UserService;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
@@ -21,6 +25,7 @@ import java.util.List;
 @Service("userService")
 public class UserServiceImpl implements UserService {
 
+    private static final Logger LOGGER = LoggerFactory.getLogger(UserServiceImpl.class);
     @Autowired
     private UserRepository userRepository;
 
@@ -38,9 +43,25 @@ public class UserServiceImpl implements UserService {
         return new UserDTO(user);
     }
 
+
     @Override
-    public void deleteUser(Long id) {
-        userRepository.delete(id);
+    public void deleteUsers(Long... ids) throws BusinessException {
+        for(Long id : ids){
+            deleteUser(id);
+        }
+    }
+
+    @Override
+    public void deleteUser(Long id) throws BusinessException{
+        //判断用户是否存在
+        Long userId = userRepository.existUserById(id);
+        if(userId != null){
+            LOGGER.debug("[user service] id 为" + id + " 用户存在，删除用户");
+            userRepository.delete(id);
+        }else {
+            LOGGER.debug("[user service] id 为" + id + "  用户不存在，抛出异常");
+            throw new BusinessException(ExceptionEnum.USER_NOT_EXIST);
+        }
     }
 
     @Override
